[WAR] Swap cmd and option keys on attached keyboard
Related iTerm2 bug: https://gitlab.com/gnachman/iterm2/-/issues/8874

Turns out that a lot of external keyboards have left and right mixed
up, making swapping left option and commands keys apply to the right
side of the keyboard.

The WAR is setting the right modifier keys to account for this.

I also swap the left modifier keys to work on keyboards that don't
have this issue, like the built-in laptop keyboard.